Revise the implementation of as_uri
